==> Code/02_conflict_heuristic_DH_compilation.ipynb <==
# Towards a Conflict Heuristic (DH 2023)

## 02. Compile Preprocessed Corpora

Last updated: 16.01.2023

julian.haeussler[at]tu-darmstadt.de

# imports

import glob
import os
import pickle
import re

# paths

path_data = r'C:\Users\Public\Data\conflict_heuristics\pickled\all'

path_results = r'C:\Users\Public\Data\conflict_heuristics\pickled'

# read in corpus lists

# Romantik

lst_files = glob.glob(os.path.join(os.getcwd(), r'C:\Users\Julian\HESSENBOX-DA\Projekte\Konflikte\Daten\Romantik (abgeschlossen)\TXT\Kernkorpus', "*.txt"))

novels_Romantik_core = []

for entry in lst_files:
    novels_Romantik_core.append(re.search(r"(?<=Kernkorpus\\)(.*)(?=.txt)",entry).group(1))

lst_files = glob.glob(os.path.join(os.getcwd(), r'C:\Users\Julian\HESSENBOX-DA\Projekte\Konflikte\Daten\Romantik (abgeschlossen)\TXT\Ergänzung I', "*.txt"))

novels_Romantik_I = []

for entry in lst_files:
    novels_Romantik_I.append(re.search(r"(?<=Ergänzung I\\)(.*)(?=.txt)",entry).group(1))

lst_files = glob.glob(os.path.join(os.getcwd(), r'C:\Users\Julian\HESSENBOX-DA\Projekte\Konflikte\Daten\Romantik (abgeschlossen)\TXT\Ergänzung II', "*.txt"))

novels_Romantik_II = []

for entry in lst_files:
    novels_Romantik_II.append(re.search(r"(?<=Ergänzung II\\)(.*)(?=.txt)",entry).group(1))

# Realismus
    
lst_files = glob.glob(os.path.join(os.getcwd(), r'C:\Users\Public\Data\Korpuserstellung\Realismus', "*.txt"))

novels_Realismus = []

for entry in lst_files:
    novels_Realismus.append(re.search(r"(?<=Realismus\\)(.*)(?=.txt)",entry).group(1))

# Naturalismus

lst_files = glob.glob(os.path.join(os.getcwd(), r'C:\Users\Public\Data\Korpuserstellung\Naturalismus', "*.txt"))

novels_Naturalismus = []

for entry in lst_files:
    novels_Naturalismus.append(re.search(r"(?<=Naturalismus\\)(.*)(?=.txt)",entry).group(1))

# create lists of lists of phrases LEMMATIZED (lst_lists_phrases_Romantik_core, lst_lists_phrases_Romantik, 
#                                              lst_lists_phrases_Realismus, lst_lists_Naturalismus)

lst_files = glob.glob(os.path.join(os.getcwd(), path_data, "*.pkl"))

lst_lists_phrases_Romantik_core_LEMMATIZED = []

for i in range(0,len(lst_files)):
    if i%4 == 0:
        name = re.search(r"(?<=all\\)(.*)(?=_phrases_lemmatized)",lst_files[i]).group(1)
        if name in novels_Romantik_core:
            with open(lst_files[i], 'rb') as f:
                lst_lists_phrases_Romantik_core_LEMMATIZED.append(pickle.load(f))

lst_lists_phrases_Romantik_LEMMATIZED = []

novels_Romantik = novels_Romantik_core.copy()
novels_Romantik.extend(novels_Romantik_I)
novels_Romantik.extend(novels_Romantik_II)

for i in range(0,len(lst_files)):
    if i%4 == 0:
        name = re.search(r"(?<=all\\)(.*)(?=_phrases_lemmatized)",lst_files[i]).group(1)
        if name in novels_Romantik:
            with open(lst_files[i], 'rb') as f:
                lst_lists_phrases_Romantik_LEMMATIZED.append(pickle.load(f))

lst_lists_phrases_Realismus_LEMMATIZED = []

for i in range(0,len(lst_files)):
    if i%4 == 0:
        name = re.search(r"(?<=all\\)(.*)(?=_phrases_lemmatized)",lst_files[i]).group(1)
        if name in novels_Realismus:
            with open(lst_files[i], 'rb') as f:
                lst_lists_phrases_Realismus_LEMMATIZED.append(pickle.load(f))

lst_lists_phrases_Naturalismus_LEMMATIZED = []

for i in range(0,len(lst_files)):
    if i%4 == 0:
        name = re.search(r"(?<=all\\)(.*)(?=_phrases_lemmatized)",lst_files[i]).group(1)
        if name in novels_Naturalismus:
            with open(lst_files[i], 'rb') as f:
                lst_lists_phrases_Naturalismus_LEMMATIZED.append(pickle.load(f))

len(lst_lists_phrases_Romantik_core_LEMMATIZED)

len(lst_lists_phrases_Romantik_LEMMATIZED)

len(lst_lists_phrases_Realismus_LEMMATIZED)

len(lst_lists_phrases_Naturalismus_LEMMATIZED)

# save

with open(path_results + '\\230116_lst_lists_phrases_Romantik_core_LEMMATIZED.pkl', 'wb') as f:
    pickle.dump(lst_lists_phrases_Romantik_core_LEMMATIZED, f)

with open(path_results + '\\230116_lst_lists_phrases_Romantik_LEMMATIZED.pkl', 'wb') as f:
    pickle.dump(lst_lists_phrases_Romantik_LEMMATIZED, f)

with open(path_results + '\\230116_lst_lists_phrases_Realismus_LEMMATIZED.pkl', 'wb') as f:
    pickle.dump(lst_lists_phrases_Realismus_LEMMATIZED, f)

with open(path_results + '\\230116_lst_lists_phrases_Naturalismus_LEMMATIZED.pkl', 'wb') as f:
    pickle.dump(lst_lists_phrases_Naturalismus_LEMMATIZED, f)

# create lists of phrases (lst_phrases_Romantik_core, lst_phrases_Romantik,
#                          lst_phrases_Realismus, lst_phrases_Naturalismus)

lst_lists_phrases_Romantik_core_TOKENIZED = []

for i in range(0,len(lst_files)):
    if i%4 == 0:
        name = re.search(r"(?<=all\\)(.*)(?=_phrases_tokenized)",lst_files[i+1]).group(1)
        if name in novels_Romantik_core:
            with open(lst_files[i+1], 'rb') as f:
                lst_lists_phrases_Romantik_core_TOKENIZED.append(pickle.load(f))

lst_lists_phrases_Romantik_TOKENIZED = []

for i in range(0,len(lst_files)):
    if i%4 == 0:
        name = re.search(r"(?<=all\\)(.*)(?=_phrases_tokenized)",lst_files[i+1]).group(1)
        if name in novels_Romantik:
            with open(lst_files[i+1], 'rb') as f:
                lst_lists_phrases_Romantik_TOKENIZED.append(pickle.load(f))

lst_lists_phrases_Realismus_TOKENIZED = []

for i in range(0,len(lst_files)):
    if i%4 == 0:
        name = re.search(r"(?<=all\\)(.*)(?=_phrases_tokenized)",lst_files[i+1]).group(1)
        if name in novels_Realismus:
            with open(lst_files[i+1], 'rb') as f:
                lst_lists_phrases_Realismus_TOKENIZED.append(pickle.load(f))

lst_lists_phrases_Naturalismus_TOKENIZED = []

for i in range(0,len(lst_files)):
    if i%4 == 0:
        name = re.search(r"(?<=all\\)(.*)(?=_phrases_tokenized)",lst_files[i+1]).group(1)
        if name in novels_Naturalismus:
            with open(lst_files[i+1], 'rb') as f:
                lst_lists_phrases_Naturalismus_TOKENIZED.append(pickle.load(f))

len(lst_lists_phrases_Romantik_core_TOKENIZED)

len(lst_lists_phrases_Romantik_TOKENIZED)

len(lst_lists_phrases_Realismus_TOKENIZED)

len(lst_lists_phrases_Naturalismus_TOKENIZED)

# save

with open(path_results + '\\230116_lst_lists_phrases_Romantik_core_TOKENIZED.pkl', 'wb') as f:
    pickle.dump(lst_lists_phrases_Romantik_core_TOKENIZED, f)

with open(path_results + '\\230116_lst_lists_phrases_Romantik_TOKENIZED.pkl', 'wb') as f:
    pickle.dump(lst_lists_phrases_Romantik_TOKENIZED, f)

with open(path_results + '\\230116_lst_lists_phrases_Realismus_TOKENIZED.pkl', 'wb') as f:
    pickle.dump(lst_lists_phrases_Realismus_TOKENIZED, f)

with open(path_results + '\\230116_lst_lists_phrases_Naturalismus_TOKENIZED.pkl', 'wb') as f:
    pickle.dump(lst_lists_phrases_Naturalismus_TOKENIZED, f)
