==> Code/03_conflict_heuristic_DH_models.ipynb <==
# Towards a Conflict Heuristic (DH 2023)

## 03. Create Models

Last updated: 16.01.2023

julian.haeussler[at]tu-darmstadt.de

# imports

from gensim.models import Word2Vec
import pickle

# paths

path_data = r'C:\Users\Public\Data\conflict_heuristics\pickled'

path_results = r'C:\Users\Public\Data\conflict_heuristics\models'

# read in corpora

with open(path_data + '\\230116_lst_lists_phrases_Romantik_LEMMATIZED.pkl', 'rb') as f:
    lst_lists_phrases_Romantik = pickle.load(f)

with open(path_data + '\\230116_lst_lists_phrases_Realismus_LEMMATIZED.pkl', 'rb') as f:
    lst_lists_phrases_Realismus = pickle.load(f)

with open(path_data + '\\230116_lst_lists_phrases_Naturalismus_LEMMATIZED.pkl', 'rb') as f:
    lst_lists_phrases_Naturalismus = pickle.load(f)

# create lists of phrases

lst_phrases_Romantik = [phrase for novel in lst_lists_phrases_Romantik for phrase in novel]

len(lst_phrases_Romantik)

lst_phrases_Romantik[:5]

lst_phrases_Realismus = [phrase for novel in lst_lists_phrases_Realismus for phrase in novel]

lst_phrases_Naturalismus = [phrase for novel in lst_lists_phrases_Naturalismus for phrase in novel]

len(lst_phrases_Realismus)

len(lst_phrases_Naturalismus)

# combine Realismus and Naturalismus

lst_phrases_RealismusNaturalismus = lst_phrases_Realismus.copy()

lst_phrases_RealismusNaturalismus.extend(lst_phrases_Naturalismus)

len(lst_phrases_RealismusNaturalismus)

# create models

# parameters: nach best practice Brottrager et al. 2022 (architecture = Skip-gram, vector size = 300, epochs = 10) 
#             + min_count = 1 (!)

model_Romantik = Word2Vec(lst_phrases_Romantik, vector_size=300, window=5, workers=6, min_count=1, sg=1, epochs=10)

model_RealismusNaturalismus = Word2Vec(lst_phrases_RealismusNaturalismus, vector_size=300, window=5, workers=6, min_count=1, sg=1, epochs=10)

# save

model_Romantik.wv.save(path_results + '\\230116_model_Romantik.kv')

model_RealismusNaturalismus.wv.save(path_results + '\\230116_model_RealismusNaturalismus.kv')

# model info

model_Romantik.wv.most_similar("natur")

model_RealismusNaturalismus.wv.most_similar("natur")

print(model_Romantik)

print(model_RealismusNaturalismus)
