==> Code/04_conflict_heuristic_DH_conflict_annotation_dictionary.ipynb <==
# Towards a Conflict Heuristic (DH 2023)

## 04. Conflict Annotation Dictionary

Last updated: 26.01.2023

julian.haeussler[at]tu-darmstadt.de

# imports

import numpy as np
import pandas as pd
import pickle

# paths

path_annotations = r"C:\Users\Julian\HESSENBOX-DA\Projekte\Konflikte\Daten\Annotation\data_prepro"

path_data = r'C:\Users\Public\Data\conflict_heuristics\pickled'

path_results = r'C:\Users\Public\Data\conflict_heuristics\csv'

path_results_pkl = r'C:\Users\Public\Data\conflict_heuristics\pickled'

#### create conflict dictionary

# read in dataframes of annotatated texts

df_eckbert = pd.read_csv(path_annotations + '\\eckbert_all_cleaned.csv')

df_eckbert.head(18)

df_judenbuche = pd.read_csv(path_annotations + '\\judenbuche_all_cleaned.csv')

df_krambambuli = pd.read_csv(path_annotations + '\\karambuli_all_cleaned.csv')

df_verwandlung = pd.read_csv(path_annotations + '\\verwandlung_all_cleaned.csv')

# create token list over all novels

lst_tokens = df_eckbert['objekt'].to_list()

lst_tokens.extend(df_judenbuche['objekt'].to_list())

lst_tokens.extend(df_krambambuli['objekt'].to_list())

lst_tokens.extend(df_verwandlung['objekt_x'].to_list())

lst_tokens[:10]

len(lst_tokens)

# create type list over all novels

lst_types = list(set(lst_tokens))

lst_types[:10]

lst_types = lst_types[1:] #remove nan

len(lst_types)

# lowercase type list

lst_types = [word.lower() for word in lst_types]

lst_types = list(set(lst_types))

len(lst_types)

lst_types[:10]

# select tags for dictionary

tags = ["emotionsbasierter_konfliktindikator", "zustandsbasierter_konfliktindikator",
        "aktionsbasierter_konfliktindikator", "konflikthinweis", "indikator_konfliktlösung",
        "indikator_konfliktintensität"]

# create dictionary

conflict_dict = {word:{tag:[] for tag in tags} for word in lst_types}

conflict_dict['ausgerenktem']

# iterate over all four novels ad add rel freq of each tag as list to dictionary

for index,row in df_eckbert.iterrows():
    l = df_eckbert.iloc[index].tolist() #token+annotations as list
    w = l[0].lower() #first item in list is token
    for tag in tags:
        v1 = l.count(tag) # count how many times a tag is in the list
        v2 = v1/6 # divided by no. of annotators
        conflict_dict[w][tag].append(v2) #add average frequency of tag to conflict dictionary

conflict_dict['ritter']

for index,row in df_judenbuche.iterrows():
    l = df_judenbuche.iloc[index].tolist()
    w = l[0].lower()
    for tag in tags:
        v1 = l.count(tag)
        v2 = v1/6
        conflict_dict[w][tag].append(v2)

for index,row in df_krambambuli.iterrows():
    l = df_krambambuli.iloc[index].tolist()
    w = l[0].lower()
    for tag in tags:
        v1 = l.count(tag)
        v2 = v1/6
        conflict_dict[w][tag].append(v2)

for index,row in df_verwandlung.iterrows():
    l = df_verwandlung.iloc[index].tolist()
    if isinstance(l[0],float) == False:
        w = l[0].lower()
        for tag in tags:
            v1 = l.count(tag)
            v2 = v1/5 #less anotators for 'Verwandlung'
            conflict_dict[w][tag].append(v2)

conflict_dict['ausgerenktem']

# compute average value for each tag for each word

conflict_dict_final = {word:{tag:0 for tag in tags} for word in lst_types}

for word in lst_types:
    for tag in tags:
        l1 = conflict_dict[word][tag]
        if len(l1) > 0:
            v = sum(l1)/len(l1)
            conflict_dict_final[word][tag]=v            

conflict_dict_final["ritter"]

# turn into df

df_conflict_final = pd.DataFrame.from_dict(conflict_dict_final,orient='index')

df_conflict_final.head()

lst_types.index("ritter")

df_conflict_final.iloc[[3120]]

# save as csv

df_conflict_final.to_csv(path_results+'\\230126_df_conflict_final.csv', encoding='utf-8-sig') 

#### get label words

# copy df

df_conflict_labels = df_conflict_final.copy()

df_conflict_labels.head()

# add column average conflict

df_conflict_labels['avrg_conflict'] = np.nan

df_conflict_labels.head()

# labels conflict

tags_conflict = ["emotionsbasierter_konfliktindikator", "zustandsbasierter_konfliktindikator",
                "aktionsbasierter_konfliktindikator", "konflikthinweis", "indikator_konfliktintensität"]

# fill column average conflict

for index,row in df_conflict_labels.iterrows():
    v=0
    for tag in tags_conflict:
        v+=df_conflict_labels.at[index,tag] #get value for specific label and add to sum of values
    df_conflict_labels.at[index,'avrg_conflict'] = v/5 # divide value by no. of labels and add to df

df_conflict_labels.at['ausgerenktem','avrg_conflict']

# get sorted list of columns

df_high_conflict = df_conflict_labels.sort_values(by=['avrg_conflict'],ascending=False).copy()

df_high_conflict.head()

labels_high_conflict = []

for index,row in df_high_conflict.iterrows():
    if df_high_conflict.at[index,'avrg_conflict'] == 0.2:
        labels_high_conflict.append(index)

labels_high_conflict

len(labels_high_conflict)

df_low_conflict = df_conflict_labels.sort_values(by=['indikator_konfliktlösung'],ascending=False).copy()

df_low_conflict.head(20)

labels_low_conflict = []

for index,row in df_low_conflict.iterrows():
    if df_low_conflict.at[index,'indikator_konfliktlösung'] > 0.4: #>0.3 -> in total 100 words
        labels_low_conflict.append(index)

labels_low_conflict

len(labels_low_conflict)

# save

with open(path_results_pkl + '\\230126_lst_labels_high_conflict.pkl', 'wb') as f:
    pickle.dump(labels_high_conflict, f)

with open(path_results_pkl + '\\230126_lst_labels_low_conflict.pkl', 'wb') as f:
    pickle.dump(labels_low_conflict, f)
