==> Code/06_conflict_heuristic_DH_filter_extreme_groups.ipynb <==
# Towards a Conflict Heuristic (DH 2023)

## 06. Filter Extreme Groups

Last updated: 25.04.2023

julian.haeussler[at]tu-darmstadt.de

# imports

import csv
import pandas as pd
import pickle

# read in files

df_VPs_Romantik_core = pd.read_csv('..//Analyseergebnisse//csv//230420_df_VPs_Romantik_core.csv', index_col=0)

df_VPs_Realismus = pd.read_csv('..//Analyseergebnisse//csv//230420_df_VPs_Realismus.csv', index_col=0)

df_VPs_Naturalismus = pd.read_csv('..//Analyseergebnisse//csv//230420_df_VPs_Naturalismus.csv', index_col=0)

# adjust columns with phrases

df_VPs_Romantik_core['phrase_tokenized'] = df_VPs_Romantik_core['phrase_tokenized'].apply(eval)

df_VPs_Romantik_core['phrase_lemmatized'] = df_VPs_Romantik_core['phrase_lemmatized'].apply(eval)

df_VPs_Realismus['phrase_tokenized'] = df_VPs_Realismus['phrase_tokenized'].apply(eval)

df_VPs_Realismus['phrase_lemmatized'] = df_VPs_Realismus['phrase_lemmatized'].apply(eval)

df_VPs_Naturalismus['phrase_tokenized'] = df_VPs_Naturalismus['phrase_tokenized'].apply(eval)

df_VPs_Naturalismus['phrase_lemmatized'] = df_VPs_Naturalismus['phrase_lemmatized'].apply(eval)

# merge dataframes

df_VPs_all = df_VPs_Romantik_core.copy()

df_VPs_all = df_VPs_all.append(df_VPs_Realismus,ignore_index=True)

df_VPs_all = df_VPs_all.append(df_VPs_Naturalismus,ignore_index=True)

len(df_VPs_all)

df_VPs_all.head()

df_VPs_all.tail()

#### Filter extreme groups

# get top 10%

i = round(len(df_VPs_all)*0.1)

print(i)

# top/bottom 10% valence ADJ

df_VPs_all_top_val_adj = df_VPs_all.sort_values('mean_val_adj',ascending=False).copy()

lst_VPs_all_top_val_adj = df_VPs_all_top_val_adj.head(i)['phrase_tokenized'].tolist()

lst_VPs_all_bottom_val_adj = df_VPs_all_top_val_adj.tail(i)['phrase_tokenized'].tolist()

# top/bottom 10% arousal ADJ

df_VPs_all_top_aro_adj = df_VPs_all.sort_values('mean_aro_adj',ascending=False).copy()

lst_VPs_all_top_aro_adj = df_VPs_all_top_aro_adj.head(i)['phrase_tokenized'].tolist()

lst_VPs_all_bottom_aro_adj = df_VPs_all_top_aro_adj.tail(i)['phrase_tokenized'].tolist()

# determine those phrases that are in 
# A top 10% arousal and top 10% valence ADJ

set_VPs_all_top_aro_adj = set((tuple(phrase) for phrase in lst_VPs_all_top_aro_adj))

set_VPs_all_top_val_adj = set((tuple(phrase) for phrase in lst_VPs_all_top_val_adj))

set_VPs_A_adj = set_VPs_all_top_aro_adj.intersection(set_VPs_all_top_val_adj)

lst_VPs_A_adj = [list(phrase) for phrase in set_VPs_A_adj]

print(len(lst_VPs_A_adj))

# B top 10% arousal and bottom 10% valence ADJ

set_VPs_all_bottom_val_adj = set((tuple(phrase) for phrase in lst_VPs_all_bottom_val_adj))

set_VPs_B_adj = set_VPs_all_top_aro_adj.intersection(set_VPs_all_bottom_val_adj)

lst_VPs_B_adj = [list(phrase) for phrase in set_VPs_B_adj]

print(len(lst_VPs_B_adj))

# C bottom 10% arousal and bottom 10% valence ADJ

set_VPs_all_bottom_aro_adj = set((tuple(phrase) for phrase in lst_VPs_all_bottom_aro_adj))

set_VPs_C_adj = set_VPs_all_bottom_aro_adj.intersection(set_VPs_all_bottom_val_adj)

lst_VPs_C_adj = [list(phrase) for phrase in set_VPs_C_adj]

print(len(lst_VPs_C_adj))

# D bottom 10% arousal and top 10% valence ADJ

set_VPs_D_adj = set_VPs_all_bottom_aro_adj.intersection(set_VPs_all_top_val_adj)

lst_VPs_D_adj = [list(phrase) for phrase in set_VPs_D_adj]

print(len(lst_VPs_D_adj))

# top/bottom 10% valence NOUN

df_VPs_all_top_val_noun = df_VPs_all.sort_values('mean_val_noun',ascending=False).copy()

lst_VPs_all_top_val_noun = df_VPs_all_top_val_noun.head(i)['phrase_tokenized'].tolist()

lst_VPs_all_bottom_val_noun = df_VPs_all_top_val_noun.tail(i)['phrase_tokenized'].tolist()

# top/bottom 10% arousal NOUN

df_VPs_all_top_aro_noun = df_VPs_all.sort_values('mean_aro_noun',ascending=False).copy()

lst_VPs_all_top_aro_noun = df_VPs_all_top_aro_noun.head(i)['phrase_tokenized'].tolist()

lst_VPs_all_bottom_aro_noun = df_VPs_all_top_aro_noun.tail(i)['phrase_tokenized'].tolist()

# determine those phrases that are in 
# A top 10% arousal and top 10% valence NOUN

set_VPs_all_top_aro_noun = set((tuple(phrase) for phrase in lst_VPs_all_top_aro_noun))

set_VPs_all_top_val_noun = set((tuple(phrase) for phrase in lst_VPs_all_top_val_noun))

set_VPs_A_noun = set_VPs_all_top_aro_noun.intersection(set_VPs_all_top_val_noun)

lst_VPs_A_noun = [list(phrase) for phrase in set_VPs_A_noun]

print(len(lst_VPs_A_noun))

# B top 10% arousal and bottom 10% valence NOUN

set_VPs_all_bottom_val_noun = set((tuple(phrase) for phrase in lst_VPs_all_bottom_val_noun))

set_VPs_B_noun = set_VPs_all_top_aro_noun.intersection(set_VPs_all_bottom_val_noun)

lst_VPs_B_noun = [list(phrase) for phrase in set_VPs_B_noun]

print(len(lst_VPs_B_noun))

# C bottom 10% arousal and bottom 10% valence NOUN

set_VPs_all_bottom_aro_noun = set((tuple(phrase) for phrase in lst_VPs_all_bottom_aro_noun))

set_VPs_C_noun = set_VPs_all_bottom_aro_noun.intersection(set_VPs_all_bottom_val_noun)

lst_VPs_C_noun = [list(phrase) for phrase in set_VPs_C_noun]

print(len(lst_VPs_C_noun))

# D bottom 10% arousal and top 10% valence NOUN

set_VPs_D_noun = set_VPs_all_bottom_aro_noun.intersection(set_VPs_all_top_val_noun)

lst_VPs_D_noun = [list(phrase) for phrase in set_VPs_D_noun]

print(len(lst_VPs_D_noun))

# top/bottom 10% conflict (Dornseiff)

df_VPs_all_top_conf_dorn = df_VPs_all.sort_values('mean_conf_dornseiff',ascending=False).copy()

lst_VPs_all_top_conf_dorn = df_VPs_all_top_conf_dorn.head(i)['phrase_tokenized'].tolist()

lst_VPs_all_bottom_conf_dorn = df_VPs_all_top_conf_dorn.tail(i)['phrase_tokenized'].tolist()

len(lst_VPs_all_top_conf_dorn)

len(lst_VPs_all_bottom_conf_dorn)

# top/bottom 10% conflict (annotation)

df_VPs_all_top_conf_anno = df_VPs_all.sort_values('mean_conf_annotation',ascending=False).copy()

lst_VPs_all_top_conf_anno = df_VPs_all_top_conf_anno.head(i)['phrase_tokenized'].tolist()

lst_VPs_all_bottom_conf_anno = df_VPs_all_top_conf_anno.tail(i)['phrase_tokenized'].tolist()

# save

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_top_val_adj.pkl', 'wb') as f:
    pickle.dump(lst_VPs_all_top_val_adj, f)
    
with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_bottom_val_adj.pkl', 'wb') as f:
    pickle.dump(lst_VPs_all_bottom_val_adj, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_top_aro_adj.pkl', 'wb') as f:
    pickle.dump(lst_VPs_all_top_aro_adj, f)
    
with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_bottom_aro_adj.pkl', 'wb') as f:
    pickle.dump(lst_VPs_all_bottom_aro_adj, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_A_adj.pkl', 'wb') as f:
    pickle.dump(lst_VPs_A_adj, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_B_adj.pkl', 'wb') as f:
    pickle.dump(lst_VPs_B_adj, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_C_adj.pkl', 'wb') as f:
    pickle.dump(lst_VPs_C_adj, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_D_adj.pkl', 'wb') as f:
    pickle.dump(lst_VPs_D_adj, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_top_val_noun.pkl', 'wb') as f:
    pickle.dump(lst_VPs_all_top_val_noun, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_bottom_val_noun.pkl', 'wb') as f:
    pickle.dump(lst_VPs_all_bottom_val_noun, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_top_aro_noun.pkl', 'wb') as f:
    pickle.dump(lst_VPs_all_top_aro_noun, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_bottom_aro_noun.pkl', 'wb') as f:
    pickle.dump(lst_VPs_all_bottom_aro_noun, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_A_noun.pkl', 'wb') as f:
    pickle.dump(lst_VPs_A_noun, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_B_noun.pkl', 'wb') as f:
    pickle.dump(lst_VPs_B_noun, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_C_noun.pkl', 'wb') as f:
    pickle.dump(lst_VPs_C_noun, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_D_noun.pkl', 'wb') as f:
    pickle.dump(lst_VPs_D_noun, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_top_conf_dorn.pkl', 'wb') as f:
    pickle.dump(lst_VPs_all_top_conf_dorn, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_bottom_conf_dorn.pkl', 'wb') as f:
    pickle.dump(lst_VPs_all_bottom_conf_dorn, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_top_conf_anno.pkl', 'wb') as f:
    pickle.dump(lst_VPs_all_top_conf_anno, f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_bottom_conf_anno.pkl', 'wb') as f:
    pickle.dump(lst_VPs_all_bottom_conf_anno, f)

#### Overlap approach 1

# overlap in top/bottom 10% -> skewed as top mostly contains label words
# top val

set_VPs_top_val_both = set_VPs_all_top_val_adj.intersection(set_VPs_all_top_val_noun)

len(set_VPs_top_val_both)

# bottom val

set_VPs_bottom_val_both = set_VPs_all_bottom_val_adj.intersection(set_VPs_all_bottom_val_noun)

len(set_VPs_bottom_val_both)

# top aro

set_VPs_top_aro_both = set_VPs_all_top_aro_adj.intersection(set_VPs_all_top_aro_noun)

len(set_VPs_top_aro_both)

# bottom aro

set_VPs_bottom_aro_both = set_VPs_all_bottom_aro_adj.intersection(set_VPs_all_bottom_aro_noun)

len(set_VPs_bottom_aro_both)
