==> Code/07_conflict_heuristic_DH_novel_ranking_average_values.ipynb <==
# Towards a Conflict Heuristic (DH 2023)

## 07. Novel Ranking Average Values

Last updated: 10.07.2023

julian.haeussler[at]tu-darmstadt.de

# imports

import csv
import glob
import math
import os
import pandas as pd
import pickle
import re

# load data

# verb phrases + values

df_VPs_Romantik_core = pd.read_csv('../Analyseergebnisse/csv/230420_df_VPs_Romantik_core.csv', index_col=0)

df_VPs_Realismus = pd.read_csv('../Analyseergebnisse/csv/230420_df_VPs_Realismus.csv', index_col=0)

df_VPs_Naturalismus = pd.read_csv('../Analyseergebnisse/csv/230420_df_VPs_Naturalismus.csv', index_col=0)

# titles

lst_titles_Romantik = []

lst_files = glob.glob(os.path.join(os.getcwd(), r'../Daten/Romantik (abgeschlossen)/TXT/Kernkorpus', '*.txt'))

for entry in lst_files:
    lst_titles_Romantik.append(re.search(r"(?<=Kernkorpus\\)(.*)(?=.txt)",entry).group(1))

lst_titles_Realismus = []

lst_files = glob.glob(os.path.join(os.getcwd(), r'../Daten/Realismus', '*.txt'))

for entry in lst_files:
    lst_titles_Realismus.append(re.search(r"(?<=Realismus\\)(.*)(?=.txt)",entry).group(1))

lst_titles_Naturalismus = []

lst_files = glob.glob(os.path.join(os.getcwd(), r'../Daten/Naturalismus', '*.txt'))

for entry in lst_files:
    lst_titles_Naturalismus.append(re.search(r"(?<=Naturalismus\\)(.*)(?=.txt)",entry).group(1))

# years

file = r'../Daten/Romantik (abgeschlossen)/Romane der Romantik_selection.xlsx'

df_meta_Romantik = pd.read_excel(file)

lst_yrs_Romantik = df_meta_Romantik['Datum Erstdruck'].tolist()

lst_yrs_Romantik[1] = 1812 # Arnim_Achim_von_Die_Kronenwaechter_Erster_Band
lst_yrs_Romantik[2] = 1817 # Arnim_Achim_von_Die_Kronenwaechter_Zweiter_Band
lst_yrs_Romantik[10] = 1795 # Hoelderlin_Friedrich_Hyperions_Jugend
lst_yrs_Romantik[11] = 1797 # Hoelderlin_Friedrich_Hyperion_oder_der_Eremit_in_Griechenland

file = r'../Daten/Metadaten_Korpus_Realismus_Romane_d-Prose.xlsx'

df_meta_Realismus = pd.read_excel(file)

lst_yrs_Realismus = df_meta_Realismus['verwendetes Datum'].tolist()

file = r'../Daten/Metadaten_Korpus_Naturalismus_Romane_d-Prose.xlsx'

df_meta_Naturalismus = pd.read_excel(file)

lst_yrs_Naturalismus = df_meta_Naturalismus['verwendetes Datum'].tolist()

# create dataframes novel ranking

df_novels_ranked_Romantik_core = pd.DataFrame(columns=['novel','year','mean_aro_adj','mean_val_adj',
                                                       'mean_conf_dornseiff', 'mean_conf_annotation'])

df_novels_ranked_Realismus_Naturalismus = pd.DataFrame(columns=['novel','year','mean_aro_adj','mean_val_adj',
                                                                'mean_conf_dornseiff', 'mean_conf_annotation'])

# fill column novels

df_novels_ranked_Romantik_core['novel'] = lst_titles_Romantik

lst_titles_Realismus_Naturalismus = lst_titles_Realismus.copy()

lst_titles_Realismus_Naturalismus.extend(lst_titles_Naturalismus)

df_novels_ranked_Realismus_Naturalismus['novel'] = lst_titles_Realismus_Naturalismus

# fill column year

df_novels_ranked_Romantik_core['year'] = lst_yrs_Romantik

lst_yrs_Realismus_Naturalismus = lst_yrs_Realismus.copy()

lst_yrs_Realismus_Naturalismus.extend(lst_yrs_Naturalismus)

df_novels_ranked_Realismus_Naturalismus['year'] = lst_yrs_Realismus_Naturalismus

# transform dataframes VPs into dicts

dict_VPs_Romantik_core = df_VPs_Romantik_core.to_dict(orient='records')

df_VPs_Realismus_Naturalismus = df_VPs_Realismus.copy()

df_VPs_Realismus_Naturalismus = df_VPs_Realismus_Naturalismus.append(df_VPs_Naturalismus, ignore_index=True)

dict_VPs_Realismus_Naturalismus = df_VPs_Realismus_Naturalismus.to_dict(orient='records')

# fill columns 'mean_aro_adj', 'mean_val_adj', 'mean_conf_dornseiff' and 'mean_conf_annotation'

# Romantik core

lst_mean_aro_adj_Romantik_core = []
lst_mean_val_adj_Romantik_core = []
lst_mean_conf_dornseiff_Romantik_core = []
lst_mean_conf_annotation_Romantik_core = []

for title in lst_titles_Romantik:
    lst_aro_all = []
    lst_val_all = []
    lst_dorn_all = []
    lst_anno_all = []
    for i in range(0,len(dict_VPs_Romantik_core)):
        if dict_VPs_Romantik_core[i]['novel_title'] == title:
            aro = dict_VPs_Romantik_core[i]['mean_aro_adj']
            val = dict_VPs_Romantik_core[i]['mean_val_adj']
            dorn = dict_VPs_Romantik_core[i]['mean_conf_dornseiff']
            anno = dict_VPs_Romantik_core[i]['mean_conf_annotation']
                
            if math.isnan(aro) == False:
                lst_aro_all.append(aro)
            if math.isnan(val) == False:
                lst_val_all.append(val)
            if math.isnan(dorn) == False:
                lst_dorn_all.append(dorn)               
            if math.isnan(anno) == False:
                lst_anno_all.append(anno)
    
    lst_mean_aro_adj_Romantik_core.append(sum(lst_aro_all)/len(lst_aro_all))
    lst_mean_val_adj_Romantik_core.append(sum(lst_val_all)/len(lst_val_all))
    lst_mean_conf_dornseiff_Romantik_core.append(sum(lst_dorn_all)/len(lst_dorn_all))
    lst_mean_conf_annotation_Romantik_core.append(sum(lst_anno_all)/len(lst_anno_all))

print(len(lst_mean_aro_adj_Romantik_core))
print(len(lst_mean_val_adj_Romantik_core))
print(len(lst_mean_conf_dornseiff_Romantik_core))
print(len(lst_mean_conf_annotation_Romantik_core))

df_novels_ranked_Romantik_core['mean_aro_adj'] = lst_mean_aro_adj_Romantik_core
df_novels_ranked_Romantik_core['mean_val_adj'] = lst_mean_val_adj_Romantik_core
df_novels_ranked_Romantik_core['mean_conf_dornseiff'] = lst_mean_conf_dornseiff_Romantik_core
df_novels_ranked_Romantik_core['mean_conf_annotation'] = lst_mean_conf_annotation_Romantik_core

df_novels_ranked_Romantik_core.head()

# Realismus Naturalismus

lst_mean_aro_adj_Realismus_Naturalismus = []
lst_mean_val_adj_Realismus_Naturalismus = []
lst_mean_conf_dornseiff_Realismus_Naturalismus = []
lst_mean_conf_annotation_Realismus_Naturalismus = []

for title in lst_titles_Realismus_Naturalismus:
    lst_aro_all = []
    lst_val_all = []
    lst_dorn_all = []
    lst_anno_all = []
    for i in range(0,len(dict_VPs_Realismus_Naturalismus)):
        if dict_VPs_Realismus_Naturalismus[i]['novel_title'] == title:
            aro = dict_VPs_Realismus_Naturalismus[i]['mean_aro_adj']
            val = dict_VPs_Realismus_Naturalismus[i]['mean_val_adj']
            dorn = dict_VPs_Realismus_Naturalismus[i]['mean_conf_dornseiff']
            anno = dict_VPs_Realismus_Naturalismus[i]['mean_conf_annotation']
                
            if math.isnan(aro) == False:
                lst_aro_all.append(aro)
            if math.isnan(val) == False:
                lst_val_all.append(val)
            if math.isnan(dorn) == False:
                lst_dorn_all.append(dorn)               
            if math.isnan(anno) == False:
                lst_anno_all.append(anno)
    
    lst_mean_aro_adj_Realismus_Naturalismus.append(sum(lst_aro_all)/len(lst_aro_all))
    lst_mean_val_adj_Realismus_Naturalismus.append(sum(lst_val_all)/len(lst_val_all))
    lst_mean_conf_dornseiff_Realismus_Naturalismus.append(sum(lst_dorn_all)/len(lst_dorn_all))
    lst_mean_conf_annotation_Realismus_Naturalismus.append(sum(lst_anno_all)/len(lst_anno_all))

print(len(lst_mean_aro_adj_Realismus_Naturalismus))
print(len(lst_mean_val_adj_Realismus_Naturalismus))
print(len(lst_mean_conf_dornseiff_Realismus_Naturalismus))
print(len(lst_mean_conf_annotation_Realismus_Naturalismus))

df_novels_ranked_Realismus_Naturalismus['mean_aro_adj'] = lst_mean_aro_adj_Realismus_Naturalismus
df_novels_ranked_Realismus_Naturalismus['mean_val_adj'] = lst_mean_val_adj_Realismus_Naturalismus
df_novels_ranked_Realismus_Naturalismus['mean_conf_dornseiff'] = lst_mean_conf_dornseiff_Realismus_Naturalismus
df_novels_ranked_Realismus_Naturalismus['mean_conf_annotation'] = lst_mean_conf_annotation_Realismus_Naturalismus

df_novels_ranked_Realismus_Naturalismus.head()

# save

df_novels_ranked_Romantik_core.to_csv('../Analyseergebnisse/csv/230710_df_novels_Romantik_core_average_values.csv', encoding='utf-8-sig') 

df_novels_ranked_Realismus_Naturalismus.to_csv('../Analyseergebnisse/csv/230710_df_novels_Realismus_Naturalismus_average_values.csv', encoding='utf-8-sig')
