==> Code/08_conflict_heuristic_DH_novel_ranking_top_percentage.ipynb <==
# Towards a Conflict Heuristic (DH 2023)

## 08. Novel Ranking Top Percentage

Last updated: 25.04.2023

julian.haeussler[at]tu-darmstadt.de

# imports

import csv
import glob
import os
import pandas as pd
import pickle
import re

# read in lists of lists of VPs

with open('..//Analyseergebnisse//pickled//230116_lst_lists_phrases_Romantik_core_TOKENIZED.pkl', 'rb') as f:
    lst_lists_phrases_Romantik = pickle.load(f)

with open('..//Analyseergebnisse//pickled//230116_lst_lists_phrases_Realismus_TOKENIZED.pkl', 'rb') as f:
    lst_lists_phrases_Realismus = pickle.load(f)

with open('..//Analyseergebnisse//pickled//230116_lst_lists_phrases_Naturalismus_TOKENIZED.pkl', 'rb') as f:
    lst_lists_phrases_Naturalismus = pickle.load(f)

len(lst_lists_phrases_Romantik)

len(lst_lists_phrases_Realismus)

len(lst_lists_phrases_Naturalismus)

# merge three lists

lst_lists_phrases_all = lst_lists_phrases_Romantik.copy()

lst_lists_phrases_all.extend(lst_lists_phrases_Realismus)

lst_lists_phrases_all.extend(lst_lists_phrases_Naturalismus)

len(lst_lists_phrases_all)

# read in lists of extreme groups

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_A_adj.pkl', 'rb') as f:
    lst_VPs_A_adj = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_A_noun.pkl', 'rb') as f:
    lst_VPs_A_noun = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_bottom_aro_adj.pkl', 'rb') as f:
    lst_VPs_bottom_aro_adj = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_bottom_aro_noun.pkl', 'rb') as f:
    lst_VPs_bottom_aro_noun = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_bottom_conf_anno.pkl', 'rb') as f:
    lst_VPs_bottom_conf_anno = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_bottom_conf_dorn.pkl', 'rb') as f:
    lst_VPs_bottom_conf_dorn = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_bottom_val_adj.pkl', 'rb') as f:
    lst_VPs_bottom_val_adj = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_bottom_val_noun.pkl', 'rb') as f:
    lst_VPs_bottom_val_noun = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_top_aro_adj.pkl', 'rb') as f:
    lst_VPs_top_aro_adj = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_top_aro_noun.pkl', 'rb') as f:
    lst_VPs_top_aro_noun = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_top_conf_anno.pkl', 'rb') as f:
    lst_VPs_top_conf_anno = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_top_conf_dorn.pkl', 'rb') as f:
    lst_VPs_top_conf_dorn = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_top_val_adj.pkl', 'rb') as f:
    lst_VPs_top_val_adj = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_all_top_val_noun.pkl', 'rb') as f:
    lst_VPs_top_val_noun = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_B_adj.pkl', 'rb') as f:
    lst_VPs_B_adj = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_B_noun.pkl', 'rb') as f:
    lst_VPs_B_noun = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_C_adj.pkl', 'rb') as f:
    lst_VPs_C_adj = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_C_noun.pkl', 'rb') as f:
    lst_VPs_C_noun = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_D_adj.pkl', 'rb') as f:
    lst_VPs_D_adj = pickle.load(f)

with open('..//Analyseergebnisse//pickled//groups//230425_lst_VPs_D_noun.pkl', 'rb') as f:
    lst_VPs_D_noun = pickle.load(f)

# create empty novel ranking

df_novels = pd.DataFrame(columns=['novel','year',
                                  'A_adj_abs','A_adj_rel','A_noun_abs','A_noun_rel',
                                  'B_adj_abs','B_adj_rel','B_noun_abs','B_noun_rel',
                                  'C_adj_abs','C_adj_rel','C_noun_abs','C_noun_rel',
                                  'D_adj_abs','D_adj_rel','D_noun_abs','D_noun_rel',
                                  'val_adj_top_abs', 'val_adj_top_rel', 'val_adj_bottom_abs', 'val_adj_bottom_rel',
                                  'val_noun_top_abs', 'val_noun_top_rel', 'val_noun_bottom_abs', 'val_noun_bottom_rel',                                
                                  'aro_adj_top_abs', 'aro_adj_top_rel', 'aro_adj_bottom_abs', 'aro_adj_bottom_rel',                                  
                                  'aro_noun_top_abs', 'aro_noun_top_rel', 'aro_noun_bottom_abs', 'aro_noun_bottom_rel', 
                                  'conf_dorn_top_abs','conf_dorn_top_rel','conf_dorn_bottom_abs','conf_dorn_bottom_rel',
                                  'conf_anno_top_abs','conf_anno_top_rel','conf_anno_bottom_abs','conf_anno_bottom_rel'])

# add novel titles

romane_titel = []

fileList = glob.glob(os.path.join(os.getcwd(), r"C:\Users\Julian\HESSENBOX-DA\Projekte\Konflikte\Daten\Romantik (abgeschlossen)\TXT\Kernkorpus", "*.txt"))

for entry in fileList:
    romane_titel.append(re.search(r"(?<=Kernkorpus\\)(.*)(?=.txt)",entry).group(1))

fileList = glob.glob(os.path.join(os.getcwd(), r"C:\Users\Public\Data\Korpuserstellung\Realismus", "*.txt"))

for entry in fileList:
    romane_titel.append(re.search(r"(?<=Realismus\\)(.*)(?=.txt)",entry).group(1))

fileList = glob.glob(os.path.join(os.getcwd(), r"C:\Users\Public\Data\Korpuserstellung\Naturalismus", "*.txt"))

for entry in fileList:
    romane_titel.append(re.search(r"(?<=Naturalismus\\)(.*)(?=.txt)",entry).group(1))

len(romane_titel)

df_novels['novel'] = romane_titel

df_novels.head()

# add publication year

# ROMANTIK

file = r'C:\Users\Julian\HESSENBOX-DA\Projekte\Konflikte\Daten\Romantik (abgeschlossen)\Romane der Romantik_selection.xlsx'

df_meta_Romantik = pd.read_excel(file)

lst_yrs_Romantik = df_meta_Romantik['Datum Erstdruck'].tolist()

lst_yrs_Romantik[:5]

# REALISMUS

file = r'C:\Users\Public\Data\Korpuserstellung\Metadaten_Korpus_Realismus_Romane_d-Prose.xlsx'

df_meta_Realismus = pd.read_excel(file)

lst_yrs_Realismus = df_meta_Realismus['verwendetes Datum'].tolist()

lst_yrs_Realismus[:5]

# NATURALISMUS

file = r'C:\Users\Public\Data\Korpuserstellung\Metadaten_Korpus_Naturalismus_Romane_d-Prose.xlsx'

df_meta_Naturalismus = pd.read_excel(file)

lst_yrs_Naturalismus = df_meta_Naturalismus['verwendetes Datum'].tolist()

lst_yrs_Naturalismus[:5]

# merge years and add to df

lst_yrs = lst_yrs_Romantik.copy()

lst_yrs.extend(lst_yrs_Realismus)

lst_yrs.extend(lst_yrs_Naturalismus)

len(lst_yrs)

df_novels['year'] = lst_yrs

df_novels.head()

df_novels.tail()

# fill novel ranking

set_VPs_A_adj = set((tuple(phrase) for phrase in lst_VPs_A_adj))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_A_adj))
    rel_freq = count/len(novel)
    df_novels.at[i,'A_adj_abs'] = count
    df_novels.at[i,'A_adj_rel'] = rel_freq

df_novels.head()

set_VPs_A_noun = set((tuple(phrase) for phrase in lst_VPs_A_noun))


for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_A_noun))
    rel_freq = count/len(novel)
    df_novels.at[i,'A_noun_abs'] = count
    df_novels.at[i,'A_noun_rel'] = rel_freq

set_VPs_B_adj = set((tuple(phrase) for phrase in lst_VPs_B_adj))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_B_adj))
    rel_freq = count/len(novel)
    df_novels.at[i,'B_adj_abs'] = count
    df_novels.at[i,'B_adj_rel'] = rel_freq

set_VPs_B_noun = set((tuple(phrase) for phrase in lst_VPs_B_noun))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_B_noun))
    rel_freq = count/len(novel)
    df_novels.at[i,'B_noun_abs'] = count
    df_novels.at[i,'B_noun_rel'] = rel_freq

set_VPs_C_adj = set((tuple(phrase) for phrase in lst_VPs_C_adj))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_C_adj))
    rel_freq = count/len(novel)
    df_novels.at[i,'C_adj_abs'] = count
    df_novels.at[i,'C_adj_rel'] = rel_freq

set_VPs_C_noun = set((tuple(phrase) for phrase in lst_VPs_C_noun))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_C_noun))
    rel_freq = count/len(novel)
    df_novels.at[i,'C_noun_abs'] = count
    df_novels.at[i,'C_noun_rel'] = rel_freq

set_VPs_D_adj = set((tuple(phrase) for phrase in lst_VPs_D_adj))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_D_adj))
    rel_freq = count/len(novel)
    df_novels.at[i,'D_adj_abs'] = count
    df_novels.at[i,'D_adj_rel'] = rel_freq

set_VPs_D_noun = set((tuple(phrase) for phrase in lst_VPs_D_noun))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_D_noun))
    rel_freq = count/len(novel)
    df_novels.at[i,'D_noun_abs'] = count
    df_novels.at[i,'D_noun_rel'] = rel_freq

df_novels.head()

df_novels['C_adj_abs'].head()

df_novels['C_adj_rel'].head()

set_VPs_top_val_adj = set((tuple(phrase) for phrase in lst_VPs_top_val_adj))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_top_val_adj))
    rel_freq = count/len(novel)
    df_novels.at[i,'val_adj_top_abs'] = count
    df_novels.at[i,'val_adj_top_rel'] = rel_freq

set_VPs_bottom_val_adj = set((tuple(phrase) for phrase in lst_VPs_bottom_val_adj))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_bottom_val_adj))
    rel_freq = count/len(novel)
    df_novels.at[i,'val_adj_bottom_abs'] = count
    df_novels.at[i,'val_adj_bottom_rel'] = rel_freq

set_VPs_top_val_noun = set((tuple(phrase) for phrase in lst_VPs_top_val_noun))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_top_val_noun))
    rel_freq = count/len(novel)
    df_novels.at[i,'val_noun_top_abs'] = count
    df_novels.at[i,'val_noun_top_rel'] = rel_freq

set_VPs_bottom_val_noun = set((tuple(phrase) for phrase in lst_VPs_bottom_val_noun))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_bottom_val_noun))
    rel_freq = count/len(novel)
    df_novels.at[i,'val_noun_bottom_abs'] = count
    df_novels.at[i,'val_noun_bottom_rel'] = rel_freq

set_VPs_top_aro_adj = set((tuple(phrase) for phrase in lst_VPs_top_aro_adj))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_top_aro_adj))
    rel_freq = count/len(novel)
    df_novels.at[i,'aro_adj_top_abs'] = count
    df_novels.at[i,'aro_adj_top_rel'] = rel_freq

set_VPs_bottom_aro_adj = set((tuple(phrase) for phrase in lst_VPs_bottom_aro_adj))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_bottom_aro_adj))
    rel_freq = count/len(novel)
    df_novels.at[i,'aro_adj_bottom_abs'] = count
    df_novels.at[i,'aro_adj_bottom_rel'] = rel_freq

set_VPs_top_aro_noun = set((tuple(phrase) for phrase in lst_VPs_top_aro_noun))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_top_aro_noun))
    rel_freq = count/len(novel)
    df_novels.at[i,'aro_noun_top_abs'] = count
    df_novels.at[i,'aro_noun_top_rel'] = rel_freq

set_VPs_bottom_aro_noun = set((tuple(phrase) for phrase in lst_VPs_bottom_aro_noun))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_bottom_aro_noun))
    rel_freq = count/len(novel)
    df_novels.at[i,'aro_noun_bottom_abs'] = count
    df_novels.at[i,'aro_noun_bottom_rel'] = rel_freq

df_novels.head()

set_VPs_top_conf_dorn = set((tuple(phrase) for phrase in lst_VPs_top_conf_dorn))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_top_conf_dorn))
    rel_freq = count/len(novel)
    df_novels.at[i,'conf_dorn_top_abs'] = count
    df_novels.at[i,'conf_dorn_top_rel'] = rel_freq

set_VPs_bottom_conf_dorn = set((tuple(phrase) for phrase in lst_VPs_bottom_conf_dorn))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_bottom_conf_dorn))
    rel_freq = count/len(novel)
    df_novels.at[i,'conf_dorn_bottom_abs'] = count
    df_novels.at[i,'conf_dorn_bottom_rel'] = rel_freq

set_VPs_top_conf_anno = set((tuple(phrase) for phrase in lst_VPs_top_conf_anno))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_top_conf_anno))
    rel_freq = count/len(novel)
    df_novels.at[i,'conf_anno_top_abs'] = count
    df_novels.at[i,'conf_anno_top_rel'] = rel_freq

set_VPs_bottom_conf_anno = set((tuple(phrase) for phrase in lst_VPs_bottom_conf_anno))

for i in range(0,len(lst_lists_phrases_all)):
    count = 0
    novel = lst_lists_phrases_all[i]
    novel_set = set((tuple(phrase) for phrase in novel))
    count = len(novel_set.intersection(set_VPs_bottom_conf_anno))
    rel_freq = count/len(novel)
    df_novels.at[i,'conf_anno_bottom_abs'] = count
    df_novels.at[i,'conf_anno_bottom_rel'] = rel_freq

df_novels.head()

df_novels.tail()

# save

df_novels.to_csv('..//Analyseergebnisse//230425_df_novel_ranking_romanticism_realism_naturalism_WSETS.csv', encoding='utf-8-sig') 

df_novels.to_pickle('..//Analyseergebnisse//230425_df_novel_ranking_romanticism_realism_naturalism_WSETS.pkl')
